rockchip-gpt-img: use `${KERNEL_DEVICETREE}` to locate `*.dtb` file
Since oe-core commit
https://github.com/openembedded/openembedded-core/commit/1860d9d3c62e2e94cd68a809385873ffd8270b6d
meta-rockchip's `create_rk_image()` function failed because it could
not locate the DeviceTree binary.  That oe-core commit disabled the
symlinks which were expected to be present.

However those symlinks were not necessary; it it a lot easier to just
use the `${KERNEL_DEVICETREE}` variable which contains the file name
within the `deploy` directory; however, we need to strip off the
directory name.  The path name in `${KERNEL_DEVICETREE}` is relative
to the `deploy/images` directory, however there is no BitBake variable
specifying this directory; there is `${DEPLOY_DIR}` which points to
`deploy`, and `${DEPLOY_DIR_IMAGE}` points to
`deploy/images/${MACHINE}`.